Add files via upload
The aim of this quiz is to calculate a spike-triggered average and to use it in determining the relevant features within a stimulus. The data provided is an experimentally recorded set of spikes from the famous H1 motion-sensitive neuron of the fly (in this case, Calliphora vicina) from the lab of Dr Robert de Ruyter van Steveninck.